src/test/java/com/neytor/timespannersoftware/controller/CityControllerTest.java
package com.neytor.timespannersoftware.controller;

import com.neytor.timespannersoftware.model.dto.City;
import com.neytor.timespannersoftware.service.CityService;
import org.jeasy.random.EasyRandom;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.BDDMockito;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.List;
import java.util.stream.Collectors;

import static org.hamcrest.Matchers.hasSize;
import static org.junit.jupiter.api.Assertions.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(MockitoExtension.class)
public class CityControllerTest {

    private EasyRandom generator = new EasyRandom();

    @InjectMocks
    private CityController controller;

    @Mock
    CityService service;

    private MockMvc mockMvc;

    @BeforeEach
    public void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }

    @Test
    void shouldReturnCityListWhenFindAll() throws Exception {
        List<City> dtoList = generator.objects(City.class, 10).collect(Collectors.toList());

        BDDMockito.given(service.findAll()).willReturn(dtoList);

        mockMvc.perform(get("/api/v1/configuration/cities"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$", hasSize(10)));
    }

    @Test
    void shouldReturnCityWhenFinById() {
    }

    @Test
    void shouldReturnCityListWhenFindByDescription() {
    }

    @Test
    void shouldReturnCreatedCityWhenCreate() {
    }

    @Test
    void shouldReturnUpdatedCityWhenUpdate() {
    }

    @Test
    void shouldInvokeRepositoryWhenDelete() {
    }
}